RAW ADC sensor has been added

+++ include/hardware/detail/SensorsManager.h
@@ -91,15 +91,15 @@ namespace detail {
 	struct AnalogPort: public destructible {
 	    /**
 	     * return the voltage present on pin 6 of the sensor port
-	     * @return voltage reading
+	     * @return raw ADC voltage reading
 	     */
-		virtual float getPin6() const = 0;
+		virtual int getPin6() const = 0;
 
 	    /**
 	     * return the voltage present on pin 1 of the sensor port
-	     * @return voltage reading
+	     * @return raw ADC voltage reading
 	     */
-		virtual float getPin1() const = 0;
+		virtual int getPin1() const = 0;
 	};
 
 	/**
